refactor(gh): collapse scattered gh modules into one `gh/` subtree (slice 3a)
Organization-as-mistake-prevention: every GitHub-touching module now
lives under `airc-lib::gh` (one discoverable, enforceable boundary)
instead of three top-level `gh_*` files:
  gh_governor.rs        -> gh/governor.rs   (the request governor / counter)
  gh_client.rs          -> gh/client.rs     (the GhClient trait)
  gh_account_registry.rs-> gh/account_registry.rs (the gist registry store)
plus gh/mod.rs documenting the chokepoint.

Crate-root re-exports are unchanged (`airc_lib::GhClient`,
`GhAccountRegistryStore`, now also `GhBudget`/`Reservation`), so consumers
keep stable names; only internal module PATHS moved
(`crate::gh_governor::` -> `crate::gh::governor::`, external
`airc_lib::gh_client::` -> `airc_lib::gh::client::`).

cargo build + clippy -D warnings clean across airc-lib + airc-cli. (Local
airc-lib test exe is blocked by this box's Windows Application Control
policy — os error 4551 — an env quirk, not a code issue; CI runs the
tests on Linux. Logic is unchanged from the pre-move passing state.)

Next (slice 3b): split the 1558-line gh/account_registry.rs into
store/document/discovery submodules — the "smaller files" half.
